Video controller, more custom errors, less API spam, right click menu, fixes
Moved the custom error handler to a different folder and added more errors to it to prevent api spam for example if you try to enter an empty token or an invalid name

Made it so the right click menu now works in the message menu so you don't have to search for users in the member list anymore and shortened up the code a bit

Added a few animations for the user menu, if you for example press on a part that is still in development then it will just flash red without any errors, going back to before it will flash red if you enter an invalid token or username

If the user menu has a popup open it will close now when you close the user menu

You can now copy the username and it has an underline on hover just like discord does

Added the default controller on videos and gifs (since for now they're counted as the same thing)

Inproved the link regex so there are no more false positives in links (Hope there won't be any need for more work on it but there will anyway)

+++ js/setToken.js
@@ -2,8 +2,12 @@ async function setToken(token) {
     let client = new Discord.Client();
     let error = false;
     try {
+        if(!token.replace(/ /, '').length){
+            errorHandler('EMPTY-TOKEN');
+            throw('EMPTY-TOKEN')
+        }
         await client.login(token).catch(err => {
-            loginErrors(token ? err : 'EMPTY-TOKEN')
+            errorHandler(err)
             throw (err)
         });
         client.destroy();
@@ -39,8 +43,10 @@ async function setToken(token) {
         load(token);
         document.getElementById('tokenbox').style.borderColor = '#313339';
     } catch (err) {
-        // Check if they have the user settings panel open. If not, show the token popup
-        document.getElementById('tokenbox').style.borderColor = '#f00';
+        // Flash red if the token is incorrect
+        let tokenBox = document.getElementById('tokenbox');
+        tokenBox.animate(animations.flashRed);
+
         // Set the error to true so it doesn't save the token
         error = true
     }
